Update tools top bar and extract VepInputSummary
