+++ src/content/app/tools/vep/components/vep-input-summary/VepInputSummary.module.css
@@ -1,0 +1,8 @@
+.label {
+  margin-right: 10px;
+}
+
+.smallLight {
+  font-size: 12px;
+  font-weight: var(--font-weight-light);
+}
